feat(idp): add account dynamic registration route
